requirements.txt
lingua-language-detector>=2.0.0
langdetect>=1.0.9
pyyaml>=6.0
requests>=2.31.0
faster-whisper>=0.10.0
